rgeometry::algorithms::convex_hull::gift_wrapping
Function convex_hull

pub fn convex_hull<T>(pts: Vec<Point<T>>) -> Result<PolygonConvex<T>, Error>
where
    T: PolygonScalar,

Convex hull of a set of points.

Gift Wrapping algorithm for finding the smallest convex polygon which contains all the given points.

Errors

Will return an error iff the input set contains less than three distinct points.

Properties

No points from the input set will be outside the returned convex polygon.

All vertices in the convex polygon are from the input set.

Time complexity

$O(n \log h)$ as h is the size of the points on convex hull

Examples

let empty_set: Vec<Point<i32,2>> = vec![];
assert_eq!(
  convex_hull(empty_set).err(),
  Some(Error::InsufficientVertices))

let dups = vec![Point::new([0,0])].repeat(3);
assert_eq!(
  convex_hull(dups).err(),
  Some(Error::InsufficientVertices))