rgeometry::algorithms::polygonization

Function new_star_polygon

pub fn new_star_polygon<T>(
    vertices: Vec<Point<T>>,
    point: &Point<T>,
) -> Result<Polygon<T>, Error>
where
    T: PolygonScalar + TotalOrd,

Creates a star-shaped polygon by sorting vertices around a center point.

A star-shaped polygon is one where all points on the boundary are visible from the center point. This function creates such a polygon by sorting the vertices counter-clockwise around the given center point.

Arguments

vertices - The vertices of the polygon (at least 3 required)

point - The center point around which vertices are sorted

Returns

Returns Ok(polygon) if a valid star polygon can be created, or an error if:

There are fewer than 3 vertices

The center point is outside the convex hull of the vertices

All vertices are collinear

Example

use rgeometry::algorithms::polygonization::new_star_polygon;
use rgeometry::data::Point;

let vertices = vec![
  Point::new([0, 0]),
  Point::new([4, 0]),
  Point::new([4, 4]),
  Point::new([0, 4]),
];
let center = Point::new([2, 2]);
let polygon = new_star_polygon(vertices, &center).unwrap();