Struct rgeometry::data::polygon::PolygonConvex

pub struct PolygonConvex<T>(_);

Implementations

impl<T> PolygonConvex<T> where
T: PolygonScalar,

pub fn new_unchecked(poly: Polygon<T>) -> PolygonConvex<T>

$O(1)$ Assume that a polygon is convex.

Safety

The input polygon has to be strictly convex, ie. no vertices are allowed to be concave or colinear.

pub fn locate(&self, pt: &Point<T, 2>) -> PointLocation

$O(\log n)$

pub fn validate(&self) -> Result<(), Error>

$O(n \log n)$

pub fn polygon(&self) -> &Polygon<T>

$O(1)$

pub fn normalize(&self) -> PolygonConvex<BigRational> where
T: PolygonScalar + Into<BigInt>,
T::ExtendedSigned: Into<BigInt>,

impl<T> PolygonConvex<T> where
T: Bounded + PolygonScalar + SampleUniform + Copy + Into<BigInt>,

pub fn random<R: ?Sized>(n: usize, rng: &mut R) -> PolygonConvex<T> where
R: Rng,

$O(n \log n)$ Uniformly sample a random convex polygon.

The output polygon is rooted in (0,0), grows upwards, and has a height and width of T::MAX.

PolygonConvex::random(3, &mut rand::thread_rng())
Run