Struct rgeometry::data::PolygonConvex

pub struct PolygonConvex<T>(_);

Implementations

impl<T> PolygonConvex<T> where
T: PolygonScalar,

pub fn new_unchecked(poly: Polygon<T>) -> PolygonConvex<T>

Assume that a polygon is convex.

Safety

The input polygon has to be strictly convex, ie. no vertices are allowed to be concave or colinear.

Time complexity

$O(1)$

pub fn locate(&self, pt: &Point<T, 2>) -> PointLocation

Locate a point relative to a convex polygon.

Time complexity

$O(\log n)$

Examples

pub fn validate(&self) -> Result<(), Error>

Validates the following properties:

Each vertex is convex, ie. not concave or colinear.

All generate polygon properties hold true (eg. no duplicate points, no self-intersections).

Time complexity

$O(n \log n)$

pub fn float(self) -> PolygonConvex<OrderedFloat<f64>> where
T: Clone + Into<f64>,

pub fn polygon(&self) -> &Polygon<T>

$O(1)$

pub fn random<R: ?Sized>(n: usize, rng: &mut R) -> PolygonConvex<T> where
T: Bounded + PolygonScalar + SampleUniform + Copy,
R: Rng,

Uniformly sample a random convex polygon.

The output polygon is rooted in (0,0), grows upwards, and has a height and width of T::max_value().

Time complexity

$O(n \log n)$

Examples

PolygonConvex::random(3, &mut rand::thread_rng())
Run

impl PolygonConvex<OrderedFloat<f64>>

pub fn normalize(&self) -> PolygonConvex<OrderedFloat<f64>>