Struct rgeometry::data::polygon::PolygonConvex

pub struct PolygonConvex<T, P = ()>(_);

Implementations

impl<T, P> PolygonConvex<T, P> where
T: PolygonScalar,

pub fn new_unchecked(poly: Polygon<T, P>) -> PolygonConvex<T, P>

O(1) Assume that a polygon is convex.

Safety

The input polygon has to be strictly convex, ie. no vertices are allowed to be concave or colinear.

pub fn locate(&self, pt: &Point<T, 2>) -> PointLocation

$O(log n)$

pub fn validate(&self) -> Result<(), Error>

pub fn polygon(&self) -> &Polygon<T, P>

impl PolygonConvex<BigRational>

pub fn random<R: ?Sized>(
n: usize,
max: usize,
rng: &mut R
) -> PolygonConvex<BigRational> where
R: Rng,

PolygonConvex::random(3, 1000, &mut rand::thread_rng())
Run