rgeometry::data
Struct Point

#[repr(transparent)]
pub struct Point<T, const N: usize = 2> {
    pub array: [T; N],
}

Fields
array: [T; N]
Implementations

impl<T, const N: usize> Point<T, N>

pub const fn new(array: [T; N]) -> Point<T, N>

pub fn new_nn(array: [T; N]) -> Point<NotNan<T>, N>
where T: FloatCore,

Panics

Panics if any of the inputs are NaN.

pub fn as_vec(&self) -> &Vector<T, N>

pub fn squared_euclidean_distance<F>(&self, rhs: &Point<T, N>) -> F
where T: Clone + Into<F>, F: NumOps + Clone + Sum,

pub fn zero() -> Self
where T: Sum,

pub fn map<U, F>(&self, f: F) -> Point<U, N>
where T: Clone, F: Fn(T) -> U,

pub fn cast<U>(&self) -> Point<U, N>
where T: Clone + Into<U>,

pub fn to_float(&self) -> Point<OrderedFloat<f64>, N>
where T: Clone + Into<f64>,

impl<T: PolygonScalar> Point<T>

pub fn cmp_distance_to(&self, p: &Point<T, 2>, q: &Point<T, 2>) -> Ordering

pub fn orient( p1: &Point<T, 2>, p2: &Point<T, 2>, p3: &Point<T, 2>, ) -> Orientation

Determine the direction you have to turn if you walk from p1 to p2 to p3.

For fixed-precision types (i8,i16,i32,i64,etc), this function is guaranteed to work for any input and never cause any arithmetic overflows.

Examples

let p1 = Point::new([ 0, 0 ]);
let p2 = Point::new([ 0, 1 ]); // One unit above p1.
// (0,0) -> (0,1) -> (0,2) == Orientation::CoLinear
assert!(Point::orient(&p1, &p2, &Point::new([ 0, 2 ])).is_colinear());
// (0,0) -> (0,1) -> (-1,2) == Orientation::CounterClockWise
assert!(Point::orient(&p1, &p2, &Point::new([ -1, 2 ])).is_ccw());
// (0,0) -> (0,1) -> (1,2) == Orientation::ClockWise
assert!(Point::orient(&p1, &p2, &Point::new([ 1, 2 ])).is_cw());

pub fn orient_along_direction( p1: &Point<T, 2>, direction: Direction<'_, T, 2>, p2: &Point<T, 2>, ) -> Orientation

pub fn orient_along_vector( p1: &Point<T, 2>, vector: &Vector<T, 2>, p2: &Point<T, 2>, ) -> Orientation

pub fn orient_along_perp_vector( p1: &Point<T, 2>, vector: &Vector<T, 2>, p2: &Point<T, 2>, ) -> Orientation

pub fn all_colinear(pts: &[Point<T>]) -> bool

pub fn ccw_cmp_around(&self, p: &Point<T, 2>, q: &Point<T, 2>) -> Ordering

Docs?

pub fn ccw_cmp_around_with( &self, z: &Vector<T, 2>, p: &Point<T, 2>, q: &Point<T, 2>, ) -> Ordering

impl<T> Point<T, 1>

pub fn x_coord(&self) -> &T

impl<T> Point<T, 2>

pub fn x_coord(&self) -> &T

pub fn y_coord(&self) -> &T

impl<T> Point<T, 3>

pub fn x_coord(&self) -> &T

pub fn y_coord(&self) -> &T

pub fn z_coord(&self) -> &T